requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
pydantic[email]==2.7.4
python-dotenv==1.0.1
supabase==2.5.0
pygame==2.5.2
stegano==0.11.2
pycryptodome==3.20.0
Pillow==10.3.0
requests==2.32.3
passlib[bcrypt]==1.7.4
python-socketio[asgi]==5.11.2
bidict==0.23.1
psycopg2-binary==2.9.9
httpx==0.27.0
jwt==1.3.1
opencv-python-headless==4.10.0.84
numpy==1.26.4
Flask-Cors==4.0.1
